fix(Plugins): Import correctly a file when using adddoc plugin
