+++ data/resources/plugins/adddoc.py
@@ -8,7 +8,9 @@
 # Description: Plugin for exporting items to CSV
 """
 
+import os
 
+from gi.repository import Gtk
 from gi.repository import GObject
 from gi.repository import Peas
 
@@ -33,9 +35,35 @@ class MiAZAddDocumentPlugin(GObject.GObject, Peas.Activatable):
 
     def add_menuitem(self, *args):
         if self.app.get_widget('workspace-menu-in-add-document') is None:
-            actions = self.app.get_service('actions')
             factory = self.app.get_service('factory')
             menu_add = self.app.get_widget('workspace-menu-in-add')
-            menuitem = factory.create_menuitem('add_docs', '... document(s)', actions.import_file, None, [])
+            menuitem = factory.create_menuitem('add_docs', '... document(s)', self.import_file, None, [])
             self.app.add_widget('workspace-menu-in-add-document', menuitem)
             menu_add.append_item(menuitem)
+
+    def import_file(self, *args):
+        factory = self.app.get_service('factory')
+        srvutl = self.app.get_service('util')
+        srvrepo = self.app.get_service('repo')
+        def filechooser_response(dialog, response, data):
+            if response == Gtk.ResponseType.ACCEPT:
+                content_area = dialog.get_content_area()
+                box = content_area.get_first_child()
+                filechooser = box.get_first_child()
+                gfile = filechooser.get_file()
+                if gfile is not None:
+                    source = gfile.get_path()
+                    btarget = srvutl.filename_normalize(source)
+                    target = os.path.join(srvrepo.docs, btarget)
+                    srvutl.filename_import(source, target)
+            dialog.destroy()
+
+        window = self.app.get_widget('window')
+        filechooser = factory.create_filechooser(
+                    parent=window,
+                    title=_('Import a single file'),
+                    target = 'FILE',
+                    callback = filechooser_response,
+                    data = None
+                    )
+        filechooser.show()
